Generalise Hamiltonian derivative calls for RHMC case
