+++ src/riemannian/integrator.jl
@@ -1,3 +1,6 @@
+import AdvancedHMC: ∂H∂θ, ∂H∂r, DualValue, PhasePoint, phasepoint, step
+using AdvancedHMC: TYPEDEF, TYPEDFIELDS, AbstractScalarOrVec, AbstractLeapfrog, step_size
+
 """
 $(TYPEDEF)
